# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/user-metadata": {
      "put": {
        "operationId": "UserMetadataController_upsertByUserId",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserMetadataDto"
                }
              }
            }
          }
        },
        "tags": [
          "User Metadata"
        ]
      },
      "post": {
        "operationId": "UserMetadataController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserMetadataDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateUserMetadataDto"
                }
              }
            }
          }
        },
        "tags": [
          "User Metadata"
        ]
      },
      "get": {
        "operationId": "UserMetadataController_findOne",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserMetadataDto"
                }
              }
            }
          }
        },
        "tags": [
          "User Metadata"
        ]
      },
      "patch": {
        "operationId": "UserMetadataController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserMetadataDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateUserMetadataDto"
                }
              }
            }
          }
        },
        "tags": [
          "User Metadata"
        ]
      },
      "delete": {
        "operationId": "UserMetadataController_remove",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "User Metadata"
        ]
      }
    },
    "/user-metadata/all": {
      "get": {
        "operationId": "UserMetadataController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedUserMetadataDto"
                }
              }
            }
          }
        },
        "tags": [
          "User Metadata"
        ]
      }
    },
    "/user-metadata/role": {
      "put": {
        "operationId": "UserMetadataController_updateRole",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateUserMetadataDto"
                }
              }
            }
          }
        },
        "tags": [
          "User Metadata"
        ]
      }
    }
  },
  "info": {
    "title": "NGX service-user-metadata API",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "UserMetadataDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "uuid": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "regular",
              "publisher",
              "admin"
            ]
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "default": "No description provided"
          },
          "__v": {
            "type": "number"
          }
        },
        "required": [
          "_id",
          "uuid",
          "role",
          "__v"
        ]
      },
      "CreateUserMetadataDto": {
        "type": "object",
        "properties": {
          "uuid": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "regular",
              "publisher",
              "admin"
            ]
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "default": "No description provided"
          }
        },
        "required": [
          "uuid"
        ]
      },
      "Object": {
        "type": "object",
        "properties": {}
      },
      "PaginatedUserMetadataDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserMetadataDto"
            }
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          }
        },
        "required": [
          "data",
          "total",
          "page",
          "limit",
          "totalPages"
        ]
      },
      "UpdateUserMetadataDto": {
        "type": "object",
        "properties": {
          "uuid": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "regular",
              "publisher",
              "admin"
            ]
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "default": "No description provided"
          }
        }
      }
    }
  }
}